Make Team Name show as "Team Name (Captain Name)"
relies on implementation detail that in the database the default team name is the captain's name. I've seen that to be true, but I'm not sure if it's something that will always stay true.

An alternative would be to check for substring match rather than exact match with something like `_series.home_team_name.indexOf(_series.home_team_captain_name) !== -1`. There may also be some flag I don't know about in the data for if the team has been named.